// team/src/patterns.enum.ts
export enum Patterns {
  CREATE_TEAM = 'TEAM_CREATE_TEAM',
  GET_TEAM = 'TEAM_GET_TEAM_BY_ID',
  UPDATE_TEAM = 'TEAM_UPDATE_TEAM',
  ADD_MEMBER = 'TEAM_ADD_MEMBER_TO_TEAM',
  REMOVE_MEMBER = 'TEAM_REMOVE_MEMBER_FROM_TEAM',
  ADD_MODERATOR = 'TEAM_MAKE_MODERATOR_MEMBER',
  REMOVE_MODERATOR = 'TEAM_REMOVE_MODERATOR_MEMBER',
  ADD_TODO = 'TEAM_ADD_TODO',
  REMOVE_TODO = 'TEAM_REMOVE_TOD',
  GET_USER_BY_ID = 'USER_GET_USER_BY_ID',
  GET_TODO_BY_ID = 'TODO_GET_TODO_BY_ID',
  ADD_USER_TEAM = 'USER_ADD_TEAM',
  REMOVE_USER_TEAM = 'USER_REMOVE_TEAM',
}